Refactor API endpoints and LLM config
