+++ chatbot/src/api/endpoints/__init__.py
@@ -1,0 +1,3 @@
+"""
+Endpoints de la API del chatbot
+"""
